Work on tests

+++ src/uwtools/tests/drivers/test_fv3.py
@@ -16,7 +16,7 @@ from pytest import fixture
 from uwtools.drivers import driver, fv3
 from uwtools.logging import log
 from uwtools.scheduler import Slurm
-from uwtools.tests.support import logged
+from uwtools.tests.support import logged, regex_logged
 
 # Fixtures
 
@@ -150,15 +150,22 @@ def test_FV3_files_copied_and_linked(config, cycle, key, task, test, tmp_path):
     assert all(getattr(dst, test)() for dst in [atm_dst, sfc_dst])
 
 
-def test_FV3_model_configure(driverobj):
+@pytest.mark.parametrize("base_file_exists", [True, False])
+def test_FV3_model_configure(base_file_exists, caplog, driverobj):
+    log.setLevel(logging.DEBUG)
     src = driverobj._rundir / "model_configure.in"
-    with open(src, "w", encoding="utf-8") as f:
-        yaml.dump({}, f)
+    if base_file_exists:
+        with open(src, "w", encoding="utf-8") as f:
+            yaml.dump({}, f)
     dst = driverobj._rundir / "model_configure"
     assert not dst.is_file()
     driverobj._driver_config["model_configure"] = {"base_file": src}
     driverobj.model_configure()
-    assert dst.is_file()
+    if base_file_exists:
+        assert dst.is_file()
+    else:
+        assert not dst.is_file()
+        assert regex_logged(caplog, f"{src}: State: Not Ready (external asset)")
 
 
 def test_FV3_namelist_file(caplog, driverobj):
